+++ packages/modules/web_themes/koala/source/src/components/charts/dailyTotals/DailyTotalsRow.vue
@@ -2,7 +2,11 @@
   <div class="component-container">
     <!-- Icon -->
     <div class="col-icon">
-      <BaseIcon :type="props.item.id" :color="props.item.color" class="base-icon" />
+      <BaseIcon
+        :type="props.item.id"
+        :color="props.item.color"
+        class="base-icon"
+      />
     </div>
     <!-- Title -->
     <div class="col-title text-weight-bold">
@@ -87,7 +91,7 @@
 <script setup lang="ts">
 import { computed } from 'vue';
 import { useMqttStore } from 'src/stores/mqtt-store';
-import BaseIcon from 'src/components/icons/BaseIcon.vue';
+import BaseIcon from 'src/components/svgIcons/BaseIcon.vue';
 import type { DailyTotalsItem } from 'src/components/models/daily-totals-model';
 
 const mqttStore = useMqttStore();
@@ -164,7 +168,7 @@ const arrowDirection = (id: string) => {
   color: var(--q-diagram-icon);
 }
 
-:deep(.base-icon svg)  {
+:deep(.base-icon svg) {
   width: 100%;
   height: 100%;
   fill: currentColor;
