+++ packages/modules/web_themes/koala/source/src/components/svgIcons/ChargePointIcon.vue
