+++ packages/modules/web_themes/koala/source/src/components/svgIcons/VehicleIcon.vue
